Move billing context providers to routes
