+++ jsapp/js/account/usage/usage.component.tsx
@@ -1,5 +1,5 @@
 import {when} from 'mobx';
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useContext, useEffect, useMemo, useState} from 'react';
 import {useLocation} from 'react-router-dom';
 import type {AccountLimit, LimitAmount} from 'js/account/stripe.types';
 import {Limits} from 'js/account/stripe.types';
@@ -13,8 +13,8 @@ import envStore from 'js/envStore';
 import {formatDate} from 'js/utils';
 import styles from './usage.module.scss';
 import useWhenStripeIsEnabled from 'js/hooks/useWhenStripeIsEnabled.hook';
-import {ProductsContext, useProducts} from '../useProducts.hook';
-import {UsageContext, useUsage} from 'js/account/usage/useUsage.hook';
+import {ProductsContext} from '../useProducts.hook';
+import {UsageContext} from 'js/account/usage/useUsage.hook';
 import moment from 'moment';
 import {YourPlan} from 'js/account/usage/yourPlan.component';
 import cx from 'classnames';
@@ -30,8 +30,8 @@ interface LimitState {
 }
 
 export default function Usage() {
-  const productsContext = useProducts();
-  const usage = useUsage();
+  const productsContext = useContext(ProductsContext);
+  const usage = useContext(UsageContext);
 
   const [limits, setLimits] = useState<LimitState>({
     storageByteLimit: Limits.unlimited,
@@ -118,81 +118,77 @@ export default function Usage() {
   }
 
   return (
-    <UsageContext.Provider value={usage}>
-      <ProductsContext.Provider value={productsContext}>
-        <div className={styles.root}>
-          <LimitNotifications accountPage />
-          <header className={styles.header}>
-            <h2 className={styles.headerText}>{t('Your usage')}</h2>
-            {typeof usage.lastUpdated === 'string' && (
-              <p className={styles.updated}>
-                {t('Last update: ##LAST_UPDATE_TIME##').replace(
-                  '##LAST_UPDATE_TIME##',
-                  usage.lastUpdated
-                )}
-              </p>
+    <div className={styles.root}>
+      <LimitNotifications accountPage />
+      <header className={styles.header}>
+        <h2 className={styles.headerText}>{t('Your usage')}</h2>
+        {typeof usage.lastUpdated === 'string' && (
+          <p className={styles.updated}>
+            {t('Last update: ##LAST_UPDATE_TIME##').replace(
+              '##LAST_UPDATE_TIME##',
+              usage.lastUpdated
             )}
-          </header>
-          {limits.stripeEnabled && <YourPlan />}
-          <div className={styles.row}>
-            <div className={cx(styles.row, styles.subrow)}>
-              <div className={styles.box}>
-                <span>
-                  <strong className={styles.title}>{t('Submissions')}</strong>
-                  <time className={styles.date}>{dateRange}</time>
-                </span>
-                <UsageContainer
-                  usage={usage.submissions}
-                  limit={limits.submissionLimit}
-                  period={usage.trackingPeriod}
-                />
-              </div>
-              <div className={styles.box}>
-                <span>
-                  <strong className={styles.title}>{t('Storage')}</strong>
-                  <div className={styles.date}>{t('per account')}</div>
-                </span>
-                <UsageContainer
-                  usage={usage.storage}
-                  limit={limits.storageByteLimit}
-                  period={usage.trackingPeriod}
-                  label={t('Total')}
-                  type={USAGE_CONTAINER_TYPE.STORAGE}
-                />
-              </div>
-            </div>
-            <div className={cx(styles.row, styles.subrow)}>
-              <div className={styles.box}>
-                <span>
-                  <strong className={styles.title}>
-                    {t('Transcription minutes')}
-                  </strong>
-                  <time className={styles.date}>{dateRange}</time>
-                </span>
-                <UsageContainer
-                  usage={usage.transcriptionMinutes}
-                  limit={limits.nlpMinuteLimit}
-                  period={usage.trackingPeriod}
-                  type={USAGE_CONTAINER_TYPE.TRANSCRIPTION}
-                />
-              </div>
-              <div className={styles.box}>
-                <span>
-                  <strong className={styles.title}>
-                    {t('Translation characters')}
-                  </strong>
-                  <time className={styles.date}>{dateRange}</time>
-                </span>
-                <UsageContainer
-                  usage={usage.translationChars}
-                  limit={limits.nlpCharacterLimit}
-                  period={usage.trackingPeriod}
-                />
-              </div>
-            </div>
+          </p>
+        )}
+      </header>
+      {limits.stripeEnabled && <YourPlan />}
+      <div className={styles.row}>
+        <div className={cx(styles.row, styles.subrow)}>
+          <div className={styles.box}>
+            <span>
+              <strong className={styles.title}>{t('Submissions')}</strong>
+              <time className={styles.date}>{dateRange}</time>
+            </span>
+            <UsageContainer
+              usage={usage.submissions}
+              limit={limits.submissionLimit}
+              period={usage.trackingPeriod}
+            />
+          </div>
+          <div className={styles.box}>
+            <span>
+              <strong className={styles.title}>{t('Storage')}</strong>
+              <div className={styles.date}>{t('per account')}</div>
+            </span>
+            <UsageContainer
+              usage={usage.storage}
+              limit={limits.storageByteLimit}
+              period={usage.trackingPeriod}
+              label={t('Total')}
+              type={USAGE_CONTAINER_TYPE.STORAGE}
+            />
           </div>
         </div>
-      </ProductsContext.Provider>
-    </UsageContext.Provider>
+        <div className={cx(styles.row, styles.subrow)}>
+          <div className={styles.box}>
+            <span>
+              <strong className={styles.title}>
+                {t('Transcription minutes')}
+              </strong>
+              <time className={styles.date}>{dateRange}</time>
+            </span>
+            <UsageContainer
+              usage={usage.transcriptionMinutes}
+              limit={limits.nlpMinuteLimit}
+              period={usage.trackingPeriod}
+              type={USAGE_CONTAINER_TYPE.TRANSCRIPTION}
+            />
+          </div>
+          <div className={styles.box}>
+            <span>
+              <strong className={styles.title}>
+                {t('Translation characters')}
+              </strong>
+              <time className={styles.date}>{dateRange}</time>
+            </span>
+            <UsageContainer
+              usage={usage.translationChars}
+              limit={limits.nlpCharacterLimit}
+              period={usage.trackingPeriod}
+            />
+          </div>
+        </div>
+      </div>
+    </div>
   );
 }
